move validation into separate folder
